Refactoirng. Move statistics providers files.
