+++ Task_02/Statistics/Abstract/HeroUsagesStatisticWinStreakProvider.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Task_02.Printers.Abstract;
 
-namespace Task_02.Statistics
+namespace Task_02.Statistics.Abstract
 {
     internal class HeroUsagesStatisticWinStreakProvider : AbstractStatisticProvider
     {
