+++ Task_02/Statistics/Abstract/NumberOfMatches/HeroUsagesStatisticMostUnfavoriteProvider.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Task_02.Printers.Abstract.NumberOfMatches;
 
-namespace Task_02.Statistics
+namespace Task_02.Statistics.Abstract.NumberOfMatches
 {
     internal class HeroUsagesStatisticMostUnfavoriteProvider : AbstractStatisticProvider
     {
