+++ Task_02/Statistics/Abstract/WinRate/HeroUsagesStatisticMostSuccesfulProvider.cs
@@ -1,11 +1,11 @@
 ﻿using System.Collections.Generic;
 using Task_02.Printers.Abstract.WinRate;
 
-namespace Task_02.Statistics
+namespace Task_02.Statistics.Abstract.WinRate
 {
     internal class HeroUsagesStatisticMostSuccessfulProvider : AbstractStatisticProvider
     {
-        public HeroUsagesStatisticMostSuccessfulProvider(List<HeroScoreInfo> infos) : base (infos, new WinRateSuccessfulPrinter()) { }
+        public HeroUsagesStatisticMostSuccessfulProvider(List<HeroScoreInfo> infos) : base(infos, new WinRateSuccessfulPrinter()) { }
 
         internal override void Show()
         {
